+++ wgpu_shadertoy/__init__.py
@@ -1,5 +1,6 @@
 from .inputs import ShadertoyChannel, ShadertoyChannelTexture
-from .shadertoy import BufferRenderPass, Shadertoy
+from .passes import BufferRenderPass, RenderPass
+from .shadertoy import Shadertoy
 
 __version__ = "0.1.0"
 version_info = tuple(map(int, __version__.split(".")))
